refactor + error handling for not found
