+++ backend/src/main/java/com/hcmus/softdes/aivideocreator/api/controllers/AuthController.java
@@ -3,8 +3,8 @@ package com.hcmus.softdes.aivideocreator.api.controllers;
 import com.hcmus.softdes.aivideocreator.api.contracts.auth.AuthResponse;
 import com.hcmus.softdes.aivideocreator.api.contracts.auth.LoginRequest;
 import com.hcmus.softdes.aivideocreator.api.services.JwtUtils;
-import com.hcmus.softdes.aivideocreator.application.dto.UserDto;
-import com.hcmus.softdes.aivideocreator.application.services.UserService;
+import com.hcmus.softdes.aivideocreator.application.user.UserDto;
+import com.hcmus.softdes.aivideocreator.application.user.UserService;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.AuthenticationManager;
