+++ backend/src/main/java/com/hcmus/softdes/aivideocreator/application/common/exceptions/NotFoundException.java
@@ -1,0 +1,7 @@
+package com.hcmus.softdes.aivideocreator.application.common.exceptions;
+
+public abstract class NotFoundException extends RuntimeException {
+    public NotFoundException(String message) {
+        super(message);
+    }
+}
