+++ backend/src/main/java/com/hcmus/softdes/aivideocreator/application/user/UserNotFoundException.java
@@ -1,0 +1,9 @@
+package com.hcmus.softdes.aivideocreator.application.user;
+
+import com.hcmus.softdes.aivideocreator.application.common.exceptions.NotFoundException;
+
+public class UserNotFoundException extends NotFoundException {
+    public UserNotFoundException() {
+        super("User not found");
+    }
+}
