+++ backend/src/main/java/com/hcmus/softdes/aivideocreator/infrastructure/repositories/InMemoryUserRepository.java
@@ -1,6 +1,6 @@
 package com.hcmus.softdes.aivideocreator.infrastructure.repositories;
 
-import com.hcmus.softdes.aivideocreator.application.interfaces.repositories.UserRepository;
+import com.hcmus.softdes.aivideocreator.application.common.interfaces.repositories.UserRepository;
 import com.hcmus.softdes.aivideocreator.domain.user.User;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Repository;
